Update jansi and aesh-readline to latest versions
