+++ bom/application/pom.xml
@@ -207,8 +207,9 @@
         <testcontainers.version>1.17.6</testcontainers.version> <!-- Make sure to also update docker-java.version to match its needs -->
         <docker-java.version>3.2.13</docker-java.version> <!-- must be the version Testcontainers use -->
         <com.dajudge.kindcontainer>1.3.0</com.dajudge.kindcontainer>
-        <aesh-readline.version>2.2</aesh-readline.version>
         <aesh.version>2.7</aesh.version>
+        <aesh-readline.version>2.4</aesh-readline.version>
+        <jansi.version>2.4.0</jansi.version> <!-- Keep in sync with aesh-readline and dekorate -->
         <!-- these two artifacts needs to be compatible together -->
         <strimzi-oauth.version>0.11.0</strimzi-oauth.version>
         <strimzi-oauth.nimbus.version>9.31</strimzi-oauth.nimbus.version>
@@ -6057,6 +6058,11 @@
                 <groupId>org.aesh</groupId>
                 <artifactId>readline</artifactId>
                 <version>${aesh-readline.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.fusesource.jansi</groupId>
+                <artifactId>jansi</artifactId>
+                <version>${jansi.version}</version>
             </dependency>
             <dependency>
                 <groupId>org.aesh</groupId>
